@@ -1,11 +1,17 @@
+/*
+ * Copyright (c) 2018.
+ * @author birfincankafein
+ *
+ */
+
 package com.birfincankafein.permissionutildemo;
 
 import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
